+++ crates/cdk-common/src/wallet/mod.rs
@@ -6,7 +6,7 @@ use std::str::FromStr;
 
 use bitcoin::hashes::{sha256, Hash, HashEngine};
 use cashu::util::hex;
-use cashu::{nut00, BlindedMessage, PaymentMethod, Proof, Proofs, PublicKey};
+use cashu::{nut00, PaymentMethod, Proof, Proofs, PublicKey};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -15,6 +15,14 @@ use crate::nuts::{
     CurrencyUnit, MeltQuoteState, MintQuoteState, SecretKey, SpendingConditions, State,
 };
 use crate::{Amount, Error};
+
+pub mod saga;
+
+pub use saga::{
+    IssueSagaState, MeltOperationData, MeltSagaState, MintOperationData, OperationData,
+    ReceiveOperationData, ReceiveSagaState, SendOperationData, SendSagaState, SwapOperationData,
+    SwapSagaState, WalletSaga, WalletSagaState,
+};
 
 /// Wallet Key
 #[derive(Debug, Clone, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
@@ -549,462 +557,6 @@ impl FromStr for OperationKind {
             "melt" => Ok(OperationKind::Melt),
             _ => Err(Error::InvalidOperationKind),
         }
-    }
-}
-
-/// States specific to send saga
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SendSagaState {
-    /// Proofs selected and reserved for sending, ready to create token
-    ProofsReserved,
-    /// Token created and ready to share, proofs marked as pending spent awaiting claim
-    TokenCreated,
-    /// Rollback in progress, reclaiming proofs via swap (transient state)
-    RollingBack,
-}
-
-impl fmt::Display for SendSagaState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            SendSagaState::ProofsReserved => write!(f, "proofs_reserved"),
-            SendSagaState::TokenCreated => write!(f, "token_created"),
-            SendSagaState::RollingBack => write!(f, "rolling_back"),
-        }
-    }
-}
-
-impl FromStr for SendSagaState {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "proofs_reserved" => Ok(SendSagaState::ProofsReserved),
-            "token_created" => Ok(SendSagaState::TokenCreated),
-            "rolling_back" => Ok(SendSagaState::RollingBack),
-            _ => Err(Error::InvalidOperationState),
-        }
-    }
-}
-
-/// States specific to receive saga
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ReceiveSagaState {
-    /// Input proofs validated and stored as pending, ready to swap for new proofs
-    ProofsPending,
-    /// Swap request sent to mint, awaiting signatures for new proofs
-    SwapRequested,
-}
-
-impl fmt::Display for ReceiveSagaState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ReceiveSagaState::ProofsPending => write!(f, "proofs_pending"),
-            ReceiveSagaState::SwapRequested => write!(f, "swap_requested"),
-        }
-    }
-}
-
-impl FromStr for ReceiveSagaState {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "proofs_pending" => Ok(ReceiveSagaState::ProofsPending),
-            "swap_requested" => Ok(ReceiveSagaState::SwapRequested),
-            _ => Err(Error::InvalidOperationState),
-        }
-    }
-}
-
-/// States specific to swap saga (wallet-side)
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SwapSagaState {
-    /// Input proofs reserved, swap request prepared, ready to execute
-    ProofsReserved,
-    /// Swap request sent to mint, awaiting signatures for new proofs
-    SwapRequested,
-}
-
-impl fmt::Display for SwapSagaState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            SwapSagaState::ProofsReserved => write!(f, "proofs_reserved"),
-            SwapSagaState::SwapRequested => write!(f, "swap_requested"),
-        }
-    }
-}
-
-impl FromStr for SwapSagaState {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "proofs_reserved" => Ok(SwapSagaState::ProofsReserved),
-            "swap_requested" => Ok(SwapSagaState::SwapRequested),
-            _ => Err(Error::InvalidOperationState),
-        }
-    }
-}
-
-/// States specific to mint (issue) saga
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum IssueSagaState {
-    /// Pre-mint secrets created and counter incremented, ready to request signatures
-    SecretsPrepared,
-    /// Mint request sent to mint, awaiting signatures for new proofs
-    MintRequested,
-}
-
-impl fmt::Display for IssueSagaState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            IssueSagaState::SecretsPrepared => write!(f, "secrets_prepared"),
-            IssueSagaState::MintRequested => write!(f, "mint_requested"),
-        }
-    }
-}
-
-impl FromStr for IssueSagaState {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "secrets_prepared" => Ok(IssueSagaState::SecretsPrepared),
-            "mint_requested" => Ok(IssueSagaState::MintRequested),
-            _ => Err(Error::InvalidOperationState),
-        }
-    }
-}
-
-/// States specific to melt saga
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum MeltSagaState {
-    /// Proofs reserved and quote locked, ready to initiate payment
-    ProofsReserved,
-    /// Melt request sent to mint, Lightning payment initiated
-    MeltRequested,
-    /// Lightning payment in progress, awaiting confirmation from network
-    PaymentPending,
-}
-
-impl fmt::Display for MeltSagaState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            MeltSagaState::ProofsReserved => write!(f, "proofs_reserved"),
-            MeltSagaState::MeltRequested => write!(f, "melt_requested"),
-            MeltSagaState::PaymentPending => write!(f, "payment_pending"),
-        }
-    }
-}
-
-impl FromStr for MeltSagaState {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "proofs_reserved" => Ok(MeltSagaState::ProofsReserved),
-            "melt_requested" => Ok(MeltSagaState::MeltRequested),
-            "payment_pending" => Ok(MeltSagaState::PaymentPending),
-            _ => Err(Error::InvalidOperationState),
-        }
-    }
-}
-
-/// Wallet saga state for different operation types
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "state", rename_all = "snake_case")]
-pub enum WalletSagaState {
-    /// Send saga states
-    Send(SendSagaState),
-    /// Receive saga states
-    Receive(ReceiveSagaState),
-    /// Swap saga states
-    Swap(SwapSagaState),
-    /// Mint (issue) saga states
-    Issue(IssueSagaState),
-    /// Melt saga states
-    Melt(MeltSagaState),
-}
-
-impl WalletSagaState {
-    /// Get the operation kind
-    pub fn kind(&self) -> OperationKind {
-        match self {
-            WalletSagaState::Send(_) => OperationKind::Send,
-            WalletSagaState::Receive(_) => OperationKind::Receive,
-            WalletSagaState::Swap(_) => OperationKind::Swap,
-            WalletSagaState::Issue(_) => OperationKind::Mint,
-            WalletSagaState::Melt(_) => OperationKind::Melt,
-        }
-    }
-
-    /// Get string representation of the inner state
-    pub fn state_str(&self) -> &'static str {
-        match self {
-            WalletSagaState::Send(s) => match s {
-                SendSagaState::ProofsReserved => "proofs_reserved",
-                SendSagaState::TokenCreated => "token_created",
-                SendSagaState::RollingBack => "rolling_back",
-            },
-            WalletSagaState::Receive(s) => match s {
-                ReceiveSagaState::ProofsPending => "proofs_pending",
-                ReceiveSagaState::SwapRequested => "swap_requested",
-            },
-            WalletSagaState::Swap(s) => match s {
-                SwapSagaState::ProofsReserved => "proofs_reserved",
-                SwapSagaState::SwapRequested => "swap_requested",
-            },
-            WalletSagaState::Issue(s) => match s {
-                IssueSagaState::SecretsPrepared => "secrets_prepared",
-                IssueSagaState::MintRequested => "mint_requested",
-            },
-            WalletSagaState::Melt(s) => match s {
-                MeltSagaState::ProofsReserved => "proofs_reserved",
-                MeltSagaState::MeltRequested => "melt_requested",
-                MeltSagaState::PaymentPending => "payment_pending",
-            },
-        }
-    }
-}
-
-/// Operation-specific data for Send operations
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SendOperationData {
-    /// Target amount to send
-    pub amount: Amount,
-    /// Memo for the send
-    pub memo: Option<String>,
-    /// Derivation counter start
-    pub counter_start: Option<u32>,
-    /// Derivation counter end
-    pub counter_end: Option<u32>,
-    /// Token data (when in Pending/Finalized state)
-    pub token: Option<String>,
-    /// Proofs being sent
-    pub proofs: Option<Proofs>,
-}
-
-/// Operation-specific data for Receive operations
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ReceiveOperationData {
-    /// Token to receive
-    pub token: Option<String>,
-    /// Derivation counter start
-    pub counter_start: Option<u32>,
-    /// Derivation counter end
-    pub counter_end: Option<u32>,
-    /// Amount received
-    pub amount: Option<Amount>,
-    /// Blinded messages for recovery
-    ///
-    /// Stored so that if a crash occurs after the mint accepts the swap,
-    /// we can use these to query the mint for signatures and reconstruct proofs.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub blinded_messages: Option<Vec<BlindedMessage>>,
-}
-
-/// Operation-specific data for Swap operations
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SwapOperationData {
-    /// Input amount
-    pub input_amount: Amount,
-    /// Output amount
-    pub output_amount: Amount,
-    /// Derivation counter start
-    pub counter_start: Option<u32>,
-    /// Derivation counter end
-    pub counter_end: Option<u32>,
-    /// Blinded messages for recovery
-    ///
-    /// Stored so that if a crash occurs after the mint accepts the swap,
-    /// we can use these to query the mint for signatures and reconstruct proofs.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub blinded_messages: Option<Vec<BlindedMessage>>,
-}
-
-/// Operation-specific data for Mint operations
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct MintOperationData {
-    /// Quote ID
-    pub quote_id: String,
-    /// Amount to mint
-    pub amount: Amount,
-    /// Derivation counter start
-    pub counter_start: Option<u32>,
-    /// Derivation counter end
-    pub counter_end: Option<u32>,
-    /// Blinded messages for recovery
-    ///
-    /// Stored so that if a crash occurs after the mint accepts the request,
-    /// we can use these to query the mint for signatures and reconstruct proofs.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub blinded_messages: Option<Vec<BlindedMessage>>,
-}
-
-/// Operation-specific data for Melt operations
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct MeltOperationData {
-    /// Quote ID
-    pub quote_id: String,
-    /// Amount to melt
-    pub amount: Amount,
-    /// Fee reserve
-    pub fee_reserve: Amount,
-    /// Derivation counter start
-    pub counter_start: Option<u32>,
-    /// Derivation counter end
-    pub counter_end: Option<u32>,
-    /// Change amount (if any)
-    pub change_amount: Option<Amount>,
-    /// Blinded messages for change recovery
-    ///
-    /// Stored so that if a crash occurs after the mint accepts the melt,
-    /// we can use these to query the mint for change signatures and reconstruct proofs.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub change_blinded_messages: Option<Vec<BlindedMessage>>,
-}
-
-/// Operation data enum
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "kind", content = "data", rename_all = "snake_case")]
-pub enum OperationData {
-    /// Send operation data
-    Send(SendOperationData),
-    /// Receive operation data
-    Receive(ReceiveOperationData),
-    /// Swap operation data
-    Swap(SwapOperationData),
-    /// Mint operation data
-    Mint(MintOperationData),
-    /// Melt operation data
-    Melt(MeltOperationData),
-}
-
-impl OperationData {
-    /// Get the operation kind
-    pub fn kind(&self) -> OperationKind {
-        match self {
-            OperationData::Send(_) => OperationKind::Send,
-            OperationData::Receive(_) => OperationKind::Receive,
-            OperationData::Swap(_) => OperationKind::Swap,
-            OperationData::Mint(_) => OperationKind::Mint,
-            OperationData::Melt(_) => OperationKind::Melt,
-        }
-    }
-}
-
-/// Wallet saga for crash-tolerant recovery.
-///
-/// Sagas represent in-progress wallet operations that need to survive crashes.
-/// They use **optimistic locking** via the `version` field to handle concurrent
-/// access from multiple wallet instances safely.
-///
-/// # Optimistic Locking
-///
-/// When multiple wallet instances share the same database (e.g., mobile app
-/// backgrounded while desktop app runs), they might both try to recover the
-/// same incomplete saga. Optimistic locking prevents conflicts:
-///
-/// 1. Each saga has a `version` number starting at 0
-/// 2. When updating, the database checks: `WHERE id = ? AND version = ?`
-/// 3. If the version matches, the update succeeds and `version` increments
-/// 4. If the version doesn't match, another instance modified it first
-///
-/// This is preferable to pessimistic locking (mutexes) because:
-/// - Works across process boundaries (multiple wallet instances)
-/// - No deadlock risk
-/// - No lock expiration/cleanup needed
-/// - Conflicts are rare in practice (sagas are short-lived)
-///
-/// Instance A reads saga with version=1
-/// Instance B reads saga with version=1
-/// Instance A updates successfully, version becomes 2
-/// Instance B's update fails (version mismatch) - it knows to skip
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct WalletSaga {
-    /// Unique operation ID
-    pub id: uuid::Uuid,
-    /// Operation kind (derived from state)
-    pub kind: OperationKind,
-    /// Saga state (operation-specific)
-    pub state: WalletSagaState,
-    /// Amount involved in the operation
-    pub amount: Amount,
-    /// Mint URL
-    pub mint_url: MintUrl,
-    /// Currency unit
-    pub unit: CurrencyUnit,
-    /// Quote ID (for mint/melt operations)
-    pub quote_id: Option<String>,
-    /// Creation timestamp (unix seconds)
-    pub created_at: u64,
-    /// Last update timestamp (unix seconds)
-    pub updated_at: u64,
-    /// Operation-specific data
-    pub data: OperationData,
-    /// Version number for optimistic locking.
-    ///
-    /// Incremented on each update. Used to detect concurrent modifications:
-    /// - If update succeeds: this instance "won" the race
-    /// - If update fails (version mismatch): another instance modified it
-    ///
-    /// Recovery code should treat version conflicts as "someone else handled it"
-    /// and skip to the next saga rather than retrying.
-    pub version: u32,
-}
-
-impl WalletSaga {
-    /// Create a new wallet saga.
-    ///
-    /// The saga is created with `version = 0`. Each successful update
-    /// will increment the version for optimistic locking.
-    pub fn new(
-        id: uuid::Uuid,
-        state: WalletSagaState,
-        amount: Amount,
-        mint_url: MintUrl,
-        unit: CurrencyUnit,
-        data: OperationData,
-    ) -> Self {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-
-        let quote_id = match &data {
-            OperationData::Mint(d) => Some(d.quote_id.clone()),
-            OperationData::Melt(d) => Some(d.quote_id.clone()),
-            _ => None,
-        };
-
-        Self {
-            id,
-            kind: state.kind(),
-            state,
-            amount,
-            mint_url,
-            unit,
-            quote_id,
-            created_at: now,
-            updated_at: now,
-            data,
-            version: 0,
-        }
-    }
-
-    /// Update the saga state and increment the version.
-    ///
-    /// This prepares the saga for an optimistic locking update.
-    /// The database layer will verify the previous version matches
-    /// before applying the update.
-    pub fn update_state(&mut self, state: WalletSagaState) {
-        self.state = state;
-        self.kind = state.kind();
-        self.updated_at = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        self.version += 1;
     }
 }
